+++ NEWS.md
@@ -3,6 +3,8 @@
 * made a few public methods of `gettext_reader` protected
 * introduced an Interface which `gettext_reader` replacement classes must implement
 * fixed calling `T::setlocale` with multiple locale arguments
+* improved return values of `T::bind_textdomain_codeset`
+* improved handling of unexpected arg type in `T::bind_textdomain_codeset` and `T::bindtextdomain`
 
 
 ### Release 2.0.0-beta1, 2025-7-20